requirements.txt
Flask>=3.0.3
flask-cors>=4.0.1
PyJWT>=2.12.0
bcrypt>=4.2.0
python-dotenv>=1.0.1
